refactor: LottoNumber 도메인 분리
